vhci(ude) port cleanup when attach exits abruptly
When usbip.exe was killed abruptly, a port of vhci(ude) was not
reclaimed. Thus vhci(ude) calls plugout at UDE cleanup routine
if not called. Ctrl+C in attach command does not suffer from this
problem. In that case, unplugging is called by a forwarder by
catching a signal.